Extract useLdapAttributes into composable
